Config file for Bambu build plates
EXPERIMENTAL!
For USE with Bambulab build plates:

Place the build plate in the front left corner on the magnetic surface.
Be careful with the z-offset. Maybe the nozzle could be to low and scatches the build plate.

ATTENTION: Before printing raise the Z-Offset 2-4 times with +0.05 and save config. Than slowly lower the z-offset while doing a test print.